Additional functionality and reset fix
Adds methods to extract the coordinates of a given state in the gridworld using the method get_state_coord(state_id). The method get_state_id(x,y) does the opposite; it takes in the coordinates as input and returns state ID. Close #8

Changes the reset() method to use a the originally prescribed state as the reset state if there is no user input (where the input defaults to None). A random initial state can be generated by using the keyword random as the input argument init_states. close #9